[FIX] helpdesk: access rights issue for helpdesk user
When a Helpdesk user tries to post a message on a ticket
where there is a helpdesk team defined, and if that user
does not belong to that team, an access right error is
raised because he does not have access to it. He should
be able to post a message anyway.

opw:2039256

closes odoo/enterprise#4957

Signed-off-by: Simon Goffin (sig) <sig@openerp.com>