[FIX] sign: drag-and-drop only on draggable elems
Create a new sign template by exporting a pdf. On the new template,
place a Signature field. Double click on the field title (should be
"Customer") so the text is selected, drag and drop the text, traceback.

The problem is that the element itself is not draggable but one of his
parents is, jquery-ui is not used in such cases and it is by default
handled by the generic jquery event dispatcher that doesn't define the
`ui` param.

opw-2044671

closes odoo/enterprise#4985

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>