[FIX] l10n_mx_reports_closing: Error with dates is fixed
On v12 the fields date are datetime not str, and in the next case I get
the traceback:
``TypeError: strptime() argument 1 must be str, not datetime.date``

1. Create the journal entry for close period.
2. Was added a feature that allows set that journal entry how entry for
closing period, for this, is necessary use the button ``Mark as closing
entry for the fiscal year``
3. When is press the button I get the traceback.

Now this is fixed.

Also was fixed the report call, because actually the report send the
date as str, and this must be consistent.

closes odoo/enterprise#4990

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>