[FIX] l10n_mx_edi: Rename finkok to quadrum.
Due to the revocal of the permission to be PAC finkok made an alliance
with a secondary PAC called **Quadrum** which help us to continue using
the webservice by finkok but signing the documents perfectly legal with
this new PAC.

Legal declaration and acceptation of the legal change for customers that
already have odoo.

For new customers they will be able to buy the service here:

https://transicion.cfdiquadrum.com.mx/cuentas/odoo/registrar/

This link should be added in the official documentation to help to new
users.

closes odoo/enterprise#4993

Signed-off-by: Josse Colpaert <jco@openerp.com>