[FIX] account_reports: Auto follow for statement sender
From Accounting / Sales / Customer Statement, selecting a
statement in need of action and clicking on "send email" do send the
message correctly, as it get logged in the chatter.

Before the commit: The user who clicked on "send email" is not
automatically added as follower of the thread, which could be
inpractical, because reply from the customer may go unnoticed.

After: User who click on "send email" is added as follower of the
discussion. This is an expected behavior in case of email sending,
because it helps following the partner reply.

OPW-2041448

closes odoo/enterprise#4996

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>