[FIX] sale_ebay: better tax handling in new order synchronisation
The new order synchronisation introduced in commit d36a1bf76bd used the taxes
directly from eBay instead of computing them on the fly
(causing possible discrepancies).
However it had a number of errors fixed here:
    - the tax is always included in unit price in eBay
    - the rate is a percentage
    - it needs to be computed relative to (unit price - tax amount)
It decomposes the given taxes into each level (SalesTax, Recycling Fees, etc).

Then the shipping tax works completely differently.
It's a percentage of the unit price (tax included) of the product(s),
independent of the shipping cost itself.
Contrarily to the product tax, this tax is not included in the price.
Therefore it is given by an amount, which we transform into a fixed tax.

opw 1924454