[FIX] sale_ebay: avoid singleton error on post-processed pickings
Since the order update can happen after modification of the order,
if the second picking had been created in the meantime this would cause a crash.
Technically it should be fixed in a better way by avoiding the message to be
posted again and again every time the order is sent by eBay, but it's late so
the real solution will wait a bit more.

opw 1924454

closes odoo/enterprise#5004

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>