[FIX] helpdesk: Missing ticket reference in mail
From the Helpdesk app create a ticket, select a customer with an
email address, and click "save". An email is sent using the default
template "Ticket: Reception Acknowledgment".

The email does not contain the reference to the just created ticket,
altough the template supports it.

This behavior is due to the message being generated at the same time as the
ticket, which still has an empty 'access_token' field, computed only when
required with '_portal_ensure_token'.

We cannot call private methods from the template for security
reasons, so we call 'get_portal_url' which ensure the token
to be computed.

Other solutions would require to patch the PortalMixin class to always
generate the token on creation, or patch HelpdeskTicket injecting the token
on creation. So, the current solution is more precise and less risky.

opw-2043783

closes odoo/enterprise#5030

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>