[FIX] web_studio: custom field does not keep image ratio
Open studio and edit a form view, add an image widget and close. Use
the just created widget to upload an image.

If the uploaded image height is greater than its width the preview will
appear squashed. This is not the right behavior, i.e. keeping the
original aspect ratio.

This is due to the fixed style imposed to the widget width and height
when created from studio.

Releasing one constraint, namely the width, allow us to keep the
original aspect ratio without creating dispoportions in the view.

Changing also tests to Studio>ViewEditorManager>Form accordingly.

opw-2040276

closes odoo/enterprise#5011

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>