[FIX] web_studio: x2m related are not stored
On a model, create a m2o
then, create a x2m relate to m2o.something_ids

Before this commit, the x2m was stored.
This is fine most of the time as we don't do anything
particular with the inverse value of the field
except when sale_timesheet is installed.
i.e.: on calendar.event, make an m2o to project
then a o2m to project.task

on the calendar grid, create an event.
while doing this, the orm tries to write [(6, False, [])]
in the new o2m, then the inverse of that field is called
(and the write of tasks is overwritten, to do some checks on project_id)

BUT

the relation of that o2m point to project.project (!) not calendar.event
and the compute inverse of the o2m will naturally inject
the calendar.event's id as the project_id to do its checks

After this commit, the new related o2m is not stored
We are aware that this is a change of behavior, some'd say dramatic
but the default of related field is not to be stored
So this fix just implements the default behavior of the ORM

OPW 2042666

closes odoo/enterprise#5013

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>