[FIX] l10n_mx_edi: Support sales from Website
In the next case:
- Install l10n_mx_edi and configure properly
- Install Website
- Go to Website -> Configuration -> Settings and enable "Invoicing" with
"Invoice what is ordered" and "Automatic Invoice"
- Make a sale from the shop

When the invoice will be paid, the flow return an error because the
payment cannot be generated if the invoice is not signed, and the
invoice is not signed because We have the after commit process.

With this fix, to avoid the dependency to website, when are
executed the Mexican process, verify that the field exists and that the
lines are related to one transaction. Then, in this case, do not wait
for the next commit, and the payment is generated correctly.

closes odoo/enterprise#5023

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>