[FIX] l10n_mx_edi: Cancel process in payment was fixed
In the next case:

1. Create a customer invoice (This must be signed as PPD)
2. Create a bank statement to pay the invoice 1
3. Go to the invoice 1 and unreconcile the payment
4. Cancel the payment that was generated in 2.

Without this fix, the payment is cancelled in Odoo, but not in the SAT,
this because when the payment is cancelled in 4, not have invoices
related, then, the `l10n_mx_edi_is_required()` method returns False.

Was fixed to cancel the CFDI when is cancel the record in Odoo and
the PAC status is signed, to cancel ot to_sign, that to has these stated
the payment is from Mexico.

closes odoo/enterprise#5024

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>