[FIX] account_reports: solved followups disappear from view
Create an invoice for a new partner with a combination of "Invoice
Date" and "Payment Terms" which makes the payment automatically
overdue, click on "Save" and "Validate". In the "Customer Statement"
section, with the filter set on "In Need of Action" find the report
related to the just created invoice, click on "Send email" and then
"Done".

The report will disappear from the view following a normal behavior,
but then when the user change the filter to display "With Overdue Invoices"
partners the manually solved report is not shown, which is not the
normal behavior.

This is due to the filter on "payment_next_action_date" being wrongly
applied to the partners, so that all partners with such field setted in the
future will not appear. Skipping the filter when we want to display
all partners with overdue invoices solves the problem.

Furthermore to avoid passing another parameter to "_compute_total_due",
not trivially doable because it is a special function which accept only
one parameter, i.e "self", we decided to implement the filter directly
in the data retrieval part so to ensure the correctness of the domain
and avoid complicating the source code.

Note that when an action is performed on a Customer Statement the
"payment_next_action_date" field is updated and so is normal behavior that
subsequent overdue invoices are not shown "In need of action",
because a minimum number of days, defined in the settings, has to pass
before being able to take further actions on the same partner

opw-2047051

closes odoo/enterprise#5067

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>