[FIX] web_studio: display all fields when creating a related
The ModelFieldSelector has a default option for filtering which makes
non searchable fields hidden.

This option makes sense when using the ModelFieldSelector in domains, as
fields need to be searchable but does not make sense when using it in
Studio for adding a related field.

To circumvent the fact that the fields were not displayed, `searchable`
attribute had been set on multiple places in view_editor_manager_tests.js
but this was a mistake. Simply displaying all fields makes more sense.

Task 2052473

closes odoo/enterprise#5096

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>