[FIX] web_studio: NULL value display as 0
From Odoo studio modify a form view, add a numeric field, then close
studio. Create some records leaving the new field untouched, i.e. to
zero, then use a filter to search all the records in which the above
mentioned field is equal to zero.

When the filter is applied, matching records will be searched for in the
database. However, since the field is unmodified, in the table will be
left uninitialized, so NULL, but the field widget still display such value as
0, potentially deceiving the user.

Modify the behavior of the field widget is too complex and would need
several modifications in the core, so our solution is to add a default
value directly in studio when creating numeric fields.
Unfortunately is not possible to have a fix for the fields that are
already in the database, because this would require ad-hoc sql queries,
so it will be effective for the records created after the new field has
been inserted into the view

opw-2050285

closes odoo/enterprise#5126

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>