[FIX] crm_voip: open lead or client
From the app switcher:
- Open the VOIP widget
- Click on a phone call
- Click on 'Opportunity' or 'Customer'

Nothing happens.

This happens because the action is `current`, but the app switcher is
not hidden.

There is a check to hide the app switcher, but the test is on the
`do_action` of the web client:

https://github.com/odoo/enterprise/blob/2ebed92eef906c97a0525fe74a35a7d678934664/web_enterprise/static/src/js/web_client.js#L161

Therefore, we do not call the `action_manager` directly, since it is
equivalent:

https://github.com/odoo/odoo/blob/2fc46a064d36d9de2c6debf9582976f938e3568b/addons/web/static/src/js/abstract_web_client.js#L221

opw-2043868

closes odoo/enterprise#5147

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>