[FIX] sale_subscription: upsell ratio
- Create a Monthly subscription Product A
- Invoice only on successful payment.
- Create a SO with A, confirm
=> a subscription is created with start date = SO date = today
- Upsell the subscription, keep the date as today

A new SO is created with an incorrect discount.

In this case, the discount should be zero since
`date = recurring_last_invoice`.

The computation is inconsistent: on one hand we compute the number of
days between 2 dates, and on the other hand we subtract an extra day.

This also fixes more generally what was solved with b4ce6918a9f21e48db4.

opw-2052446

closes odoo/enterprise#5153

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>