[FIX] helpdesk: create ticket from website
- Activate Website Form on the Helpdesk team settings
- Publish the team on the website
- Submit a ticket with the public user

Although the name is mandatory, the email value is saved as name.

The root cause is in `generate_recipients`, which doesn't take into
account any name for partner creation:

https://github.com/odoo/odoo/blob/d5072c2ec6d90c5760ab9ef44bee265b96970523/addons/mail/models/mail_template.py#L396

Ideally, `generate_recipients` should be able to take the name into
account, for example if the email address looks like
`"Raoul Grosbedon <r.g@grosbedon.fr>"`. However, this implies some
intrusive changes in the `mail` module for a minor issue. Therefore, we
limit the scope of the fix to `helpdesk`.

opw-2052997

closes odoo/enterprise#5171

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>