[FIX] web_studio: add group_by in search view
When you edit a search view with studio, and add a new 'group by' the
name of the 'group by' is the same as the name of the field. This can
create issues when using the context with search_default.

For example, in CRM search view, if you add a group by costumers, and go
to a contact and click to the opportunities button. The group by that
you just created will show by default.

opw-2048484

closes odoo/enterprise#5179

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>