[FIX] l10n_nl_reports: VAT return=>no double abroad
In d5a08e76a3 (july 2019) the sign of the VAT return report about sections

- 4A Supplies/services from non-EU countries (VAT)
- 4B Supplies/services from EU countries (VAT)

was kept instead of using `abs` as was done since 5e7d91d5ea (may 2016).

This was intended to fix the case of credit note and purchase having the
same sign, but the report is defined like this:

- section 5A is sum of sales taxes (section 1, 2 and 4)
- section 5B contains sum of taxes to claim
- section 5C is section 5A minus section 5B

Thus before july 2019:

- for a purchase: tax to pay in 4A/4B would balance tax to claim in 5B
- for a credit note: tax to pay in 4A/4B would double tax to claim in 5B

After july 2019:

- for a purchase: tax to pay in 4A/4B would double tax to claim in 5B
- for a credit note: tax to pay in 4A/4B would double tax to claim in 5B

So this is now wrong in any case and not just credit note.

With this changeset, we inverse sign to have as expected:

- for a purchase: tax to pay in 4A/4B would balance tax to claim in 5B
- for a credit note: tax to pay in 4A/4B would balance tax to claim in 5B

It makes sense to inverse the rates since they were negative tax rates
to start with (and it is what was done in 8.0).

example with numbers:

if we have vendor bill 1000€ with "Inkopen import binnen EU hoog" taxes,
we have -210€ VAT to bear, and 210€ VAT to claim:

- taxes before d5a08e7: 4b=210€, 5a=210€, 5b=210€, 5c=0€, 5g=0€ (OK)
- taxes without fix: 4b=-210€, 5a=-210€, 5b=210€, 5c=-420€, 5g=-420€
- taxes with fix: 4b=210€, 5a=210€, 5b=210€, 5c=0€, 5g=0€ (OK)

if we refund this vendor bill before cancelling it:

- taxes before d5a08e7: 4b=210€, 5a=210€, 5b=-210€, 5c=420€, 5g=420€
- taxes without fix: 4b=210€, 5a=210€, 5b=-210€, 5c=420€, 5g=420€
- taxes with fix: 4b=-210€, 5a=-210€, 5b=-210€, 5c=0€, 5g=0€ (OK)

opw-2044238
closes #5177

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>