[FIX] mrp_workorder: block quantity update of an MO with done quality checks
Create a manufacturing order with quality checks.
Process part of the order, passing a quality check.
Update the MO's production quantity: the quality checks are not updated.

Because of the multiple cases that should be dealt with if we wanted to
correctly update the quality checks, this is a limitation of this version.
As such the only true bugfix is to block the feature.

opw 2046058

closes odoo/enterprise#5193

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>