[FIX] account_reports: 2 of mn/de/nl installed=>tl issue
When a module is installed, thanks to a92abf0be2 missing menu and action
name translations of an accounting report are copied from existing ones.

But if you are in the following configurations:

- installed account_reports
- installed l10n_nl_reports
- installed "dutch / nederlands" language
- then install either german or mongolian accounting

This causes an issue, because we will insert two ir.translation:

{"lang": "nl_NL", "type": "model", name: "ir.actions.client,name",
"res_id": {new action res_id}, "t.src": "Profit and Loss", ...}

which breaks because of the UNIQUE constraint in odoo/odoo@160d65770.

The second translations of "Profit and Loss" comes from l10n_nl_reports
since 5e98dea0.

With this changeset, when we only insert up to 1 translation per
language and per menu or action record.

opw-2055833
closes #5211

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>