[FIX] stock_barcode: Success popup shows up on cancel
Open the Barcode app, select "Operations", then "Receipts", use an
available one or create one, click on "Validate": an helper widget will
draw, close it using the "x" on the top right.

The view return to the receipts as if the action clicked was "Apply",
also showing the popup "Success" on top right. The issue is limited to a
similarity in the view as the action is not actually performed.

Fixing the check by adding further condition on the behavior of the
modal solves the issue, then top account for a difference in the
view (i.e. no modal) when the barcode scanner is used a shortcut to
the 'success' path has to be introduced

opw-2057451

closes odoo/enterprise#5262

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>