[FIX] account_invoice_extract: ignore tracebacks of invoices pre-updated
Before, when updating the status of an invoice processed by the OCR, we
were pre-updating other invoices at the same time. If an error occurred
during the processing of those, the traceback would show as if it was
was caused byg the current invoice. This behaviour was very confusing
and could prevent the user from updating the current invoice.

We are now ignoring any error happening during the update of those
pre-updated invoices.

closes odoo/enterprise#5294

Signed-off-by: Florian Daloze (fda) <fda@odoo.com>