[FIX] marketing_automation: marketing participant search
Before this commit, it wasn't possible to search or filter by
participant in the tracer list view or in the participants list view.

This occurs because the participant resource is a fields.Reference and a
search function must be created to be able to search or filter in that
particular field.

opw-2058687