[FIX] account_bank_statement_import_camt: debit amount_currency
When importing a CAMT with statements in multi-currency, if the amount is negative, the amount_currency doesn't reflect that.
Consequence: impossible to do a bank reconciliation as it shows the amount_currency in the wrong column.

This fixes the issue by copying the sign of amount to amount_currency.

opw 2035136

closes odoo/enterprise#5308

Signed-off-by: Simon Goffin (sig) <sig@openerp.com>