[FIX] account_budget: Unposted entries included in practical amount
From the accounting app create a new budgetary position for some account
in which there are posted and unposted entries. Create a budget with
that position but no Analytic account associated. Save and Confirm.

The Practical amount calculated will take into account also the unposted
journal entries, which should not be the case since by definition do not
impact the balances until posted.

Adding the proper check on the domain fix the issue

opw-2057254

closes odoo/enterprise#5313

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>