[FIX] account_ponto: fix synchronization throttle
The previous code was rejecting the synchronization if the last_refresh
was from 0 to 5 minutes ago, but also if it was from 1:00 to 1:05, 2:00
to 2:05, etc up to from 23:00 to 23:05 ago.

The new code is only rejecting it if last_refresh was from 0 to 5
minutes ago, and the syntax is more explicit.

closes odoo/enterprise#5371

Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>