[FIX] account_budget: pivot on computed field
Show the pivot view of the account budget report. Traceback.

The pivot view shows a sum of each column using a `{field}:sum`
read_group field name but as the field is not stored it is not possible
to compute it hence the traceback. As the read_group overrided in the
budjet model does not handle aggregate function, they have been removed.

closes odoo/enterprise#5373

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>