[FIX] account_ponto: avoid missing transactions
Comparing the value date and the last synchronization date is not enough
to determine when to stop a synchronization, because the bank may still
receive transactions valued at this date. In Ponto sandbox, they are not
ordered on any date, so we need to consider another criteria.

As Ponto allows us to retrieve the transactions starting from any
identifier, if we know one, ignore the last sync date and use this
mechanism to make sure that we don't miss any transaction.

By the way, the last_sync date limit now truncates the timestamp to a
date because it is compared to dates without a time part, and the latest
transaction identifier is only checked once per transaction page.

OPW-2056404

closes odoo/enterprise#5354

Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>