[IMP] l10n_mx_edi: Method that update SAT status improved
The method `l10n_mx_edi_update_sat_status()` in the payment could be
called in any PAC status whenever the record have a CFDI related
(l10n_mx_edi_cfdi field).

For example the next case:
A CFDI calls the method to cancel, but return an error with the PAC, and
the `PAC status` is `to_cancel`, In this case I need to check if the
CFDI was cancelled in the SAT, and is necessary that method.

Also was removed the `is_required()` validation, for the next reason:

A payment is related to an invoice, and was signed, this will be
cancelled, and the user unreconcile first the invoice with the payment.
In this case the method `is_required()` returns False, but the payment
has a CFDI, then must be updated the SAT status.

Based on the
[invoice](https://github.com/odoo/enterprise/commit/53cd592983a6d6939ba0b29be8897c06e615eb16)

closes odoo/enterprise#5507

X-original-commit: ea6b245cd11bd9cc178d66f39bbd56f8c45915f7
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>