[FIX] mrp_mps: 3 steps manufacturing
- Activate 'Multi-Step Routes'
- Go to Inventory > Configuration > Warehouses
- In 'Resupply', set 'Manufacture' to 'Pick components, manufacture and
  then store products (3 steps)'
- Create a product A, set the route 'Manufacture'
- Go to Manufacturing > Planning > Master Production Schedule
- On the line 'To Receive / To Supply / Produce', set 50 units
- Click on the green target button 'Launch Procurement'

The quantity is reset to zero.

This is because the 'Manufacture' route uses the 'Manufacture' rule
for which the 'WH/Post-Production' location is not under the 'WH/Stock'
location. Stock moves in 'WH/Post-Production' are excluded from the
search domain, and therefore they are not accounted for.

We extend the domain to the top-level warehouse location in order to
include all moves inside the warehouse hierarchy.

opw-2044151

closes odoo/enterprise#5497

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>