[FIX] delivery_dhl: weight not split by package on labels
- Install the DHL connector.
- In the settings, check "Operations > Delivery Packages" as well as
"Warehouse > Multi-Step Routes".
- Create a SO for two products with the delivery method "DHL USA".
- Click on "Delivery", and select your line.
- In the grid of products, click the icon next to the Done quantity.
- Generate two destination packages with their respective weight.
- Back in the delivery, click "Validate".

In the chatter, a PDF is shown containing the two labels for the delivery.

Before this commit:

the weight of each package was the sum of the weight of all the packages.

After this commit:

the labels show the weight of their own package next to the total weight.
Ex: 2/3 kg, for a package weighting 2 kg, if it's shipped with a second
package of 1 kg.

closes odoo/enterprise#5598

Opw: 2070609
X-original-commit: c4ee69170f63e0b278533e7e07d2f99d2840b4ca
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>