[FIX] sale_ebay: update each order only once in new synchronisation
Commit d36a1bf76bd introduced a new synchronisation mechanism based on orders.
The order confirmation has been put in its own function, in particular to
allow further information to be put in the Odoo order.
However, since there is a message posted (picking_ids[-1].message_post),
this happens every time an order is returned
(which happens at each update of its status on eBay).

opw 1924454

closes odoo/enterprise#5596

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>