[FIX] web_studio: edit t-field in report editor (non-debug)
Before this commit, when a `t-field` was used inside a `<p>`,
the field was not editable in non-debug mode. For example:
```xml
<p>
	<span t-field="o_my_field"/>
</p>
```
When clicking on the field, it prompts edition of the `<p>` as
a text component.

This happens due to `<p>` being considered a text component in
non-debug mode, due to not having a `t-field` on it. However,
since inner-`span` contains `t-field`, the `<p>` should not have
been considered as a text component.

This commit fixes the issue by restraining the selector for node
being considered as text components: nodes of text components must
not have any children with a `t-field`.

Note that this problem did not happen in debug mode because children
of text node were not "blacklisted" for edition.

Task-Id 2057666

closes odoo/enterprise#5578

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>