[FIX] account_bank_statement_import_csv: reconciliation interface not showing
Install accounting and account_bank_statement_import_csv, import the
test_csv.csv file located in account_bank_statement_import_csv/test_csv
The bank statement is created but will not show the reconciliation
interface correctly.

This is due to a mismatch between what the callback for the import
expect as parameters and what it is actually passed.
Renaming a parameter solve the issue.

opw-2076107

closes odoo/enterprise#5797

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>